requirements.txt
fastapi==0.122.0
pydantic==2.12.5
uvicorn[standard]==0.38.0
qdrant-client==1.16.1
openai==2.8.1
python-dotenv==1.2.1
openai-agents==0.6.1
tiktoken
